mcp-server-twitter/pyproject.toml
[project]
name = "mcp-server-twitter"
version = "0.1.0"
description = "MCP Server for Twitter integration"
authors = [
    { name = "XyAgent", email = "agent@xyber.inc" },
]
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "mcp>=1.8.0,<2.0.0",
    "pydantic-settings>=2.0",
    "uvicorn[standard]",
    "fastmcp>=2.3.4",
    "fastapi>=0.115.12",
    "tweepy>=4.14.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.8.0",
    "async-lru>=2.0.0",
    "oauthlib>=3.2.0",
    "tenacity>=8.5.0",
    "x402[evm]>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "pytest-asyncio>=0.23.2",
    "httpx>=0.25.0",
    "ruff>=0.1.6",
    "black>=23.11.0",
    "isort>=5.12.0",
    "aioresponses>=0.7.4", # For mocking async HTTP requests
]
debug = [
    "debugpy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server_twitter"]

# --- Tool Configurations ---

[tool.ruff]
line-length = 88
target-version = "py312"
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "W",   # pycodestyle warnings
    "C90", # mccabe complexity
    "N",   # pep8-naming
    "D",   # pydocstyle
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "S",   # flake8-bandit
]
ignore = [
    "D100", # Missing docstring in public module
    "D104", # Missing docstring in public package
    "D107", # Missing docstring in __init__
    "D203", # 1 blank line required before class docstring
    "D212", # Multi-line docstring summary should start at the first line
    "S101", # Use of assert detected (common in tests)
]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "build",
    "dist",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["S101", "D"]  # Allow assert in tests, skip docstrings

[tool.mypy]
python_version = "3.12"
files = ["src/**/*.py"]
exclude = ["tests/", "__pycache__/"]
strict = true
ignore_missing_imports = true
show_error_codes = true
warn_return_any = false
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=70",
    "--strict-markers",
    "--disable-warnings",
    "-v",
]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
